Implemented new methods for increaseing and decreasing a color channel.
Since the state of each LED is stored in an array of pulse widths,
I added a method to convert back to a color byte.
To increase or decrease the shade of G, R or B, the higher level methods
will first convert the value back to color byte, increment or decrement as
neccessary and convert it again to pulse widths.